Move `MappingFileReader` and `MappingDirReader`
