--- src/main/java/net/fabricmc/mappingio/format/MappingDirReader.java
@@ -1,39 +1,0 @@
-/*
- * Copyright (c) 2023 FabricMC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package net.fabricmc.mappingio.format;
-
-import java.io.IOException;
-import java.io.Reader;
-import java.nio.file.Path;
-
-import org.jetbrains.annotations.ApiStatus;
-import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
-
-import net.fabricmc.mappingio.MappingReader;
-import net.fabricmc.mappingio.MappingVisitor;
-
-@ApiStatus.Internal
-public interface MappingDirReader extends MappingReader {
-	@Override
-	default void read(Path dir, MappingVisitor visitor) throws IOException {
-		read(null, dir, visitor);
-	}
-
-	@Override
-	void read(@Nullable Reader reader, @NotNull Path dir, MappingVisitor visitor) throws IOException;
-}
