+++ src/main/java/net/fabricmc/mappingio/format/srg/TsrgFileReader.java
@@ -27,10 +27,10 @@ import java.util.Objects;
 
 import net.fabricmc.mappingio.MappedElementKind;
 import net.fabricmc.mappingio.MappingFlag;
+import net.fabricmc.mappingio.MappingReader.MappingFileReader;
 import net.fabricmc.mappingio.MappingUtil;
 import net.fabricmc.mappingio.MappingVisitor;
 import net.fabricmc.mappingio.format.ColumnFileReader;
-import net.fabricmc.mappingio.format.MappingFileReader;
 
 public final class TsrgFileReader implements MappingFileReader {
 	private TsrgFileReader() {
